New ontologies and Aadarsh's work
Created two new ontologies. Note: don't use the ones with the SD ending. That stands for Star Dog which is another tool I'm using just to visualize some stuff but it's behaving weird so am trying stuff to get it to work. The latest ontology without SD is the official one.