TBox Separation
I've been wondering for a while how to separate the schema model (what ontologists call TBox) from all the data (Abox) for a while. I was thinking of using separate graphs but I realized it isn't that complex. I just wrote a query to delete all the instances with exceptions for things I want to keep.